+++ package.json
@@ -4,6 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
+    "agenda": "^2.0.0",
     "bcrypt": "^3.0.0",
     "bluebird": "^3.5.1",
     "cors": "^2.8.4",
